Board Page about 80% done
Running into an issue with columns, have a ton of white space on the
left side of the page that is driving me insane.

Added in past SVPs thumbnails & current board images.

Removed current board image placeholder, it would be repetitive on this
page. An image like that could be used on the main page.